pyproject.toml
[tool.poetry]
name = "arxplorer"
version = "0.1.0"
description = "An advanced system for searching, retrieving, and analyzing academic papers from arXiv using AI-powered agents."
authors = ["Marco Fago <fago.marco.ai@gmail.com>"]
readme = "README.md"
packages = [{ include = "arxplorer", from = "src" }]
homepage = "https://github.com/marfago/ArXplorer"
repository = "https://github.com/marfago/ArXplorer"
documentation = "https://github.com/marfago/ArXplorer#readme"
keywords = ["arxiv", "research", "academic", "papers", "AI", "literature-review"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
license = "Apache-2.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/marfago/ArXplorer/issues"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.1"
pymupdf4llm = "^0.0.17"
arxiv = "^2.1.3"
pydantic = "^2.10.4"
arize-phoenix = "^7.5.1"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-otlp = "^1.29.0"
dash = "^2.18.2"
dash-bootstrap-components = "^1.6.0"
docling = "^2.15.1"
click = "^8.1.8"
black = "^25.1.0"
openinference-instrumentation-dspy = "^0.1.14"
transformers = "^4.48.2"
waitress = "^3.0.2"
dspy = "2.5.43"
sqlalchemy = "^2.0.38"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
requests-mock = "^1.12.1"
pytest-cov = "^4.1.0"
flake8 = "^7.1.1"
flake8-bugbear = "^24.12.12"
coverage-badge = "^1.1.0"
coveragepy-lcov = "^0.1.2"

[tool.poetry.group.torch]
optional = true

[tool.poetry.group.torch.dependencies]
torch = "^2.6.0"
torchvision = "^0.21.0"
torchaudio = "^2.6.0"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[[tool.poetry.source]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=arxplorer --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 130
target-version = ['py310', 'py311', 'py312']

[tool.coverage.run]
source = ["src"]

[tool.poetry.scripts]
arxplorer = "arxplorer.server:main"